Refactor OpenAI API key management: rename model and update client initialization
